# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Digital Wallet API documentation",
        "title": "Digital Wallet API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Abdel Rahman Hussein",
            "url": "https://github.com/abdelrahman146"
        },
        "version": "1.0"
    },
    "host": "localhost:3401",
    "basePath": "/api/v1",
    "paths": {
...
        "/backoffice/exchange-rates": {
            "get": {
                "description": "Get all exchange rates",
                "tags": [
                    "Exchange Rate"
                ],
                "summary": "Get all exchange rates",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ExchangeRate"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an exchange rate based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Rate"
                ],
                "summary": "Create a new exchange rate",
                "parameters": [
                    {
                        "description": "Create Exchange Rate Request",
                        "name": "exchangeRate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateExchangeRateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/model.ExchangeRate"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/backoffice/exchange-rates/wallets/{walletId}": {
            "get": {
                "description": "Get all exchange rates of a wallet",
                "tags": [
                    "Exchange Rate"
                ],
                "summary": "Get all exchange rates of a wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ExchangeRate"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/backoffice/exchange-rates/{exchangeRateId}": {
            "put": {
                "description": "Update an exchange rate based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Rate"
                ],
                "summary": "Update an exchange rate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exchange Rate ID",
                        "name": "exchangeRateId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Exchange Rate Request",
                        "name": "exchangeRate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/model.ExchangeRate"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an exchange rate by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Rate"
                ],
                "summary": "Delete an exchange rate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exchange Rate ID",
                        "name": "exchangeRateId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/backoffice/wallets/{walletId}/transactions/exchange": {
            "post": {
                "description": "Create an exchange transaction based on the provided request",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create an exchange transaction",
                "parameters": [
                    {
                        "description": "Create Exchange Transaction Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/service.ExchangeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.ErrorResponseBody"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.ErrorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "httpCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.ExchangeRate": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "exchangeRate": {
                    "description": "@swaggertype number",
                    "type": "number"
                },
                "fromWalletId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "minimumAmount": {
                    "type": "integer"
                },
                "tierId": {
                    "type": "string"
                },
                "toWalletId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Transaction": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "availableAmount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "expireAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/types.JSONB"
                },
                "newBalance": {
                    "type": "integer"
                },
                "previousBalance": {
                    "type": "integer"
                },
                "programId": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
...
        "service.CreateExchangeRateRequest": {
            "type": "object",
            "required": [
                "exchangeRate",
                "fromWalletId",
                "toWalletId"
            ],
            "properties": {
                "exchangeRate": {
                    "type": "number"
                },
                "fromWalletId": {
                    "type": "string"
                },
                "tierId": {
                    "type": "string"
                },
                "toWalletId": {
                    "type": "string"
                }
            }
        },
...
        "service.ExchangeResponse": {
            "type": "object",
            "properties": {
                "fromTransaction": {
                    "$ref": "#/definitions/model.Transaction"
                },
                "toTransaction": {
                    "$ref": "#/definitions/model.Transaction"
                }
            }
        },
...
        "types.JSONB": {
            "type": "object",
            "additionalProperties": true
        }
    }
}